LayoutTests/fast/js/delete-syntax-expected.txt

This test checks whether various forms of delete expression are allowed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS delete x is true
PASS delete window.x is true
PASS delete window["x"] is true
PASS delete (x) is true
PASS delete (window.x) is true
PASS delete (window["x"]) is true
PASS (y, delete x) is true
PASS delete ((x)) is true
PASS delete ((window.x)) is true
PASS delete ((window["x"])) is true
PASS delete (y, x) is true
PASS delete (true ? x : y) is true
PASS delete nonexistent is true
PASS delete window.nonexistent is true
PASS delete window["nonexistent"] is true
PASS delete (nonexistent) is true
PASS delete (window.nonexistent) is true
PASS delete (window["nonexistent"]) is true
PASS delete "x" is true
PASS delete (2 + 3) is true
PASS Math.cos is mathCos
PASS Math.sin is undefined
PASS Math.tan is null
PASS RegExp.prototype.compile is regExpPrototypeCompile
PASS RegExp.prototype.exec is undefined
PASS RegExp.prototype.test is null
PASS Object.getOwnPropertyNames(Object.prototype).indexOf('__defineSetter__') is -1
PASS successfullyParsed is true

TEST COMPLETE
